refactor : change path user repository
