+++ src/application/usecases/authentication/auth-use-case.ts
@@ -2,7 +2,7 @@ import { Hash } from "../../../domain/hash";
 import { Jwt } from "../../../domain/jwt";
 import { Logging } from "../../../domain/logging";
 import { Authentication, Login, Token } from "../../../domain/auth";
-import { UserRepository } from "../../../domain/users/users-repository";
+import { UserRepository } from "../../../domain/repositories/users-repository";
 import { AppError, InvalidCredentials, NotFound } from "../../errors/errors";
 import { HttpStatusCode } from "../../../domain/http-status";
 
@@ -11,7 +11,7 @@ export class AuthenticationUseCase implements Login {
     readonly usersRepository: UserRepository,
     readonly jwtService: Jwt,
     readonly bcrypt: Hash,
-    readonly logging: Logging,
+    readonly logging: Logging
   ) {}
   async invoke(user: Authentication): Promise<Token> {
     const existentUser = await this.usersRepository.findByEmail(user.email);
@@ -20,13 +20,13 @@ export class AuthenticationUseCase implements Login {
     }
     const isValidPassword = await this.bcrypt.compare(
       user.password,
-      existentUser.password,
+      existentUser.password
     );
     if (!isValidPassword) {
       this.logging.warn(`User credential are invalid ${user.email}`);
       throw new InvalidCredentials(
         "Invalid credentials",
-        HttpStatusCode.Unauthorized,
+        HttpStatusCode.Unauthorized
       );
     }
     try {
@@ -34,11 +34,11 @@ export class AuthenticationUseCase implements Login {
       return { token };
     } catch (error) {
       this.logging.error(
-        `Some internal server error has been ocurred trying log user : ${error}`,
+        `Some internal server error has been ocurred trying log user : ${error}`
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError,
+        HttpStatusCode.InternalServerError
       );
     }
   }
