+++ src/application/usecases/users/create-user-use-case.ts
@@ -1,6 +1,6 @@
 import { Hash } from "../../../domain/hash";
 import { Logging } from "../../../domain/logging";
-import { UserRepository } from "../../../domain/users/users-repository";
+import { UserRepository } from "../../../domain/repositories/users-repository";
 import { AppError, UserAlreadyExist } from "../../errors/errors";
 import { HttpStatusCode } from "../../../domain/http-status";
 import { User } from "../../../domain/entities/user-entity";
@@ -13,7 +13,7 @@ export class CreateUserUseCase implements CreateUser {
   constructor(
     readonly usersRepository: UserRepository,
     readonly bcrypt: Hash,
-    readonly logging: Logging,
+    readonly logging: Logging
   ) {}
   async invoke(user: User): Promise<User> {
     const existentUser = await this.usersRepository.findByEmail(user.email);
@@ -32,11 +32,11 @@ export class CreateUserUseCase implements CreateUser {
       return this.usersRepository.create(newUser);
     } catch (error) {
       this.logging.error(
-        `Some Internal server error has been ocurred trying create a new user : ${error}`,
+        `Some Internal server error has been ocurred trying create a new user : ${error}`
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError,
+        HttpStatusCode.InternalServerError
       );
     }
   }
