+++ src/presentation/routes/activity/activty-routes.ts
@@ -8,5 +8,5 @@ export const activitiesManagementRoutes = Router();
 activitiesManagementRoutes.post(
   "/",
   activityValidatorMiddleware.validate(),
-  adaptRoute(makeCreateActivityController())
+  adaptRoute(makeCreateActivityController()),
 );
